Tasks

+++ tasks/task1.py
@@ -2,8 +2,8 @@
 # Write a program that prompts the user to enter the base and 
 
 
-base = int(input("Enter the base: "))
-height = int(input("enter the height: "))
+base = float(input("Enter the base: "))
+height = float(input("enter the height: "))
 area = 0.5 * base * height
 
 print(area)
